// frontend/src/api/generated/model/ndmsTunnelKind.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */

export type NdmsTunnelKind = typeof NdmsTunnelKind[keyof typeof NdmsTunnelKind];


export const NdmsTunnelKind = {
  amnezia_wireguard: 'amnezia_wireguard',
  wireguard: 'wireguard',
  openvpn: 'openvpn',
  ike: 'ike',
  l2tp: 'l2tp',
  sstp: 'sstp',
  openconnect: 'openconnect',
  http_proxy: 'http_proxy',
  https_proxy: 'https_proxy',
  socks5_proxy: 'socks5_proxy',
} as const;